ZOOKEEPER-2869: Configurable retry policies in ClientCnxn.SendThread
When a client's connection to a server is lost, it currently retries
with a random time in the range [0, 1) seconds. This can cause excessive
retries when multiple client connects are dropped simultaneously.

Allow for configurable retry strategies to be used, including an
exponential retry policy with optional jitter, as described
[here](https://www.awsarchitectureblog.com/2015/03/backoff.html)

The semantics of random retries in the range [0, 1) is preserved by
default.